fix dependency cycle
run linter fix
