+++ packages/frontend/src/components/icons/symbols/ActiveIcon.tsx
@@ -1,4 +1,5 @@
 import React from 'react'
+
 import { Icon } from '../Icon'
 
 export function ActiveIcon() {
