+++ packages/frontend/src/components/table/props/getBridgesTableColumns.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
-import { BridgesRiskViewEntry } from '../../../pages/bridges-risk/BridgesRiskView'
-import { BridgesTvlViewEntry } from '../../../pages/bridges-tvl/BridgesTvlView'
+
+import { BridgesRiskViewEntry } from '../../../pages/bridges-risk/types'
+import { BridgesTvlViewEntry } from '../../../pages/bridges-tvl/types'
 import { TVLBreakdown } from '../../TVLBreakdown'
 import { IndexCell } from '../IndexCell'
 import { NoInfoCell } from '../NoInfoCell'
