+++ packages/frontend/src/pages/bridges-risk/types.ts
@@ -1,0 +1,15 @@
+import { ProjectRiskViewEntry } from '@l2beat/config'
+
+export interface BridgesRiskViewEntry {
+  name: string
+  slug: string
+  type: 'layer2' | 'bridge'
+  warning?: string
+  isArchived?: boolean
+  isVerified?: boolean
+  category: string
+  destination: ProjectRiskViewEntry
+  validatedBy?: ProjectRiskViewEntry
+  sourceUpgradeability?: ProjectRiskViewEntry
+  destinationToken?: ProjectRiskViewEntry
+}
