+++ packages/frontend/src/pages/bridges-tvl/types.ts
@@ -1,0 +1,21 @@
+import { ProjectRiskViewEntry } from '@l2beat/config'
+
+import { TVLBreakdownProps } from '../../components/TVLBreakdown'
+
+export interface BridgesTvlViewEntry {
+  type: 'bridge' | 'layer2'
+  name: string
+  slug: string
+  warning?: string
+  isArchived?: boolean
+  isUpcoming?: boolean
+  isVerified?: boolean
+  tvl?: string
+  tvlBreakdown?: TVLBreakdownProps
+  oneDayChange?: string
+  sevenDayChange?: string
+  bridgesMarketShare?: string
+  combinedMarketShare?: string
+  validatedBy?: ProjectRiskViewEntry
+  category: string
+}
