+++ packages/frontend/src/pages/scaling-risk/view/ScalingRiskView.tsx
@@ -1,32 +1,17 @@
-import { Layer2, ProjectRiskViewEntry } from '@l2beat/config'
 import React from 'react'
 
-import { ScalingLegend } from '../../../components/ScalingLegend'
-import { TabNavigation } from '../../../components/TabNavigation'
 import { ActiveIcon } from '../../../components/icons/symbols/ActiveIcon'
 import { ArchivedIcon } from '../../../components/icons/symbols/ArchivedIcon'
-import { RowConfig, TableView } from '../../../components/table/TableView'
+import { ScalingLegend } from '../../../components/ScalingLegend'
 import { getScalingRowProps } from '../../../components/table/props/getScalingRowProps'
 import { getScalingRiskColumns } from '../../../components/table/props/getScalingTableColumns'
+import { RowConfig, TableView } from '../../../components/table/TableView'
+import { TabNavigation } from '../../../components/TabNavigation'
+import { ScalingRiskViewEntry } from './types'
 
 export interface ScalingRiskViewProps {
   items: ScalingRiskViewEntry[]
   upcomingEnabled?: boolean
-}
-
-export interface ScalingRiskViewEntry {
-  name: string
-  slug: string
-  provider?: Layer2['technology']['provider']
-  warning?: string
-  isArchived?: boolean
-  isVerified?: boolean
-  isUpcoming?: boolean
-  stateValidation: ProjectRiskViewEntry
-  dataAvailability: ProjectRiskViewEntry
-  upgradeability: ProjectRiskViewEntry
-  sequencerFailure: ProjectRiskViewEntry
-  validatorFailure: ProjectRiskViewEntry
 }
 
 export function ScalingRiskView({ items }: ScalingRiskViewProps) {
