+++ packages/frontend/src/pages/scaling-risk/view/types.ts
@@ -1,0 +1,16 @@
+import { Layer2, ProjectRiskViewEntry } from '@l2beat/config'
+
+export interface ScalingRiskViewEntry {
+  name: string
+  slug: string
+  provider?: Layer2['technology']['provider']
+  warning?: string
+  isArchived?: boolean
+  isVerified?: boolean
+  isUpcoming?: boolean
+  stateValidation: ProjectRiskViewEntry
+  dataAvailability: ProjectRiskViewEntry
+  upgradeability: ProjectRiskViewEntry
+  sequencerFailure: ProjectRiskViewEntry
+  validatorFailure: ProjectRiskViewEntry
+}
